pyproject.toml
[project]
name = "farmsetu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "celery[redis]>=5.5.1",
    "django>=4.2.20",
    "django-celery-beat>=2.8.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
